Compressor implementation updated.
Currently, it is just a skeleton that gathers all the files to be compressed and spawns the indicated number of threads that will eventually compress in parallel those files.
There is a manager that accepts petitions from the threads to process the next file.
For now, the threads just ask the manager for the next item and prints the name of the item to be processed.