Updated Decompressor.java and added OneInputToMulti.java
The class OneInputToMulti opens the set of files created by the compressor and processes each file inside the archive.
The Decompressor should take each chunk of bytes and write them onto the output destination.
For now the decompressor just prints the name of the files it should process, but no actual files are written.