DecompressorImpl.java now write files.
The decompressor now is able to process each item and store the file in it's destination.
Clean up a bit the code to use an interface to process each chunk of demultiplexed data.
Dear Agoda-ian, it is interesting to note that the whole system (compressor and decompressor) still do not use any compression algorithm yet. :smile: